requirements.txt
requests>=2.25.1
python-telegram-bot>=20.0
configparser>=5.0.0
schedule>=1.1.0